fix: move reselector to reducer
